fix: move RawTxResult struct from internal to wdk package.
